% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_estimation.R
\name{beta_OSE}
\alias{beta_OSE}
\title{beta one-step elimination}
\usage{
beta_OSE(X, y, variables, lambda_cutoff)
}
\arguments{
\item{X}{is the regression matrix with the marginals, the fixed effects, the outcome of interest}

\item{y}{is the outcome of interest}

\item{variables}{is a vector containing all the variables we want to regress on}

\item{lambda_cutoff}{is the cutoff on beta values}
}
\value{
returns the estimated support
}
\description{
Estimate support with beta one-step elimination
}
