% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{create_dosages_from_dummies}
\alias{create_dosages_from_dummies}
\title{From dummy columns to dosages columns}
\usage{
create_dosages_from_dummies(data, arms)
}
\arguments{
\item{data}{is a dataframe with the data}

\item{arms}{is a named list, and each element of the list is a vector containing the dummy columns of one arm dosages, in the order of increasing dosages and without the dosage 0 dummy. The name of this element must be the arm name.}
}
\value{
returns the dataframe "data" with one column per arm showing the dosage.
}
\description{
Going from the dummy columns to a unique column per arm with the dosages
}
\examples{
sms_lvl_1 = c(0,0,0,1,0,1)
sms_lvl_2 = c(0,1,1,0,0,0)
incentive_lvl_1 = c(1,0,1,0,1,0)
incentive_lvl_2 = c(0,0,0,1,0,1)
information_lvl_1 = c(1,0,0,1,0,1)
information_lvl_2 = c(0,0,0,0,1,0)
information_lvl_3 = c(0,1,1,0,0,0)
arms = list(sms = c('sms_lvl_1','sms_lvl_2'),
            incentive = c('incentive_lvl_1','incentive_lvl_2'),
            info = c('information_lvl_1','information_lvl_2','information_lvl_3'))
df = data.frame(sms_lvl_1, sms_lvl_2, incentive_lvl_1, incentive_lvl_2, information_lvl_1, information_lvl_2, information_lvl_3)
create_dosages_from_dummies(data,arms)
}
