% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{do_TVA}
\alias{do_TVA}
\title{TVA Main function}
\usage{
do_TVA(
  data,
  arms,
  y,
  fes = c(),
  w = NULL,
  cutoff = NULL,
  estim_func = "pval_OSE",
  compare_to_zero = FALSE,
  clusters = NULL
)
}
\arguments{
\item{data}{is the dataframe containing all our data}

\item{arms}{is a vector containing the column names of all the arms}

\item{y}{is the column name of the outcome of interest}

\item{fes}{(optional) is a vector containing the column names of all the fixed effects, empty by default}

\item{w}{(optional) is the column name of the weights}

\item{cutoff}{(optional) is the cutoff used in the support estimation. If empty, TVA will suggest its own cutoff based on the penalty / support size tradeoff.}

\item{estim_func}{(optional) is the estimation function we should used. Possible arguments are :\cr
\enumerate{
\item 'pval_MSE': a multiple step elimination on p-values\cr
\item 'pval_OSE': a one step elimination on p-values\cr
\item 'puffer_N_LASSO': a LASSO OLS with a Puffer_N transformation\cr
\item 'beta_OSE': a one step elimination on beta values\cr
\item 'puffer_LASSO': a LASSO OLS with a Puffer transformation\cr
(2) and (3) should be equivalent, as well as (4) and (5)
}}

\item{compare_to_zero}{(optional) is a boolean, FALSE by default. \cr
If TRUE, the code considers that a policy dominates a marginal if all dosages are greater\cr
If FALSE, then they must also have the exact same activated arms (the zeros of the policy vectors are at identical indexes)}

\item{clusters}{(optional) is the column name that corresponds to the clusters in the data, that should be used in the final pooled OLS. Please refer to estimatr::lm_robust documentation for more information on this parameter.}
}
\value{
returns a list containing:\cr
\itemize{
\item data: the data with new columns giving pooling information\cr
\item marginal_support: a dataframe with all the marginals in the support and their according id\cr
\item pools_summary: a dataframe with information on each pool\cr
\item unique_policy: a dataframe with all the possible unique policies and their according pool id\cr
\item fes_support: the intersection between the estimated support and the fixed effects\cr
\item pooled_ols: the result of the final OLS on the pooled data\cr
\item winners_effect: the result of the best pooled policy effect, downsized by the winners curse algorithm
}
}
\description{
Perform the whole TVA algorithm
}
\examples{
arms = c('financial_incentive','reminder','information')
fes = c('fes_1')
y = 'outcome'
w = 'weights'
A1 = c(0,0,0,0,0,1,1,1,1,1)
A2 = c(1,1,0,0,1,1,0,0,1,1)
A3 = c(0,1,2,3,0,3,2,1,0,1)
F1 = c(0,1,0,0,0,1,0,1,0,0)
Y  = c(5,4,3,5,4,5,4,2,3,2)
W  = c(1,1,1,2,1,2,2,1,1,2)
data = data.frame(financial_incentive = A1, reminder = A2, information = A3, fes_1 = F1, outcome = Y, weights=W)
TVA(data,arms,y, fes,w,0.3,'pval_OSE',FALSE,FALSE)
}
