% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_functions.R
\name{elbow}
\alias{elbow}
\title{Find elbow}
\usage{
elbow(X, Y)
}
\arguments{
\item{X}{is a numeric vector}

\item{Y}{is a numeric vector, same size as X}
}
\value{
returns the value of X corresponding to the elbow of  the curve
}
\description{
Find the elbow of a set of points.
}
