% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{fill_marginals_matrix}
\alias{fill_marginals_matrix}
\title{Fill marginals matrix}
\usage{
fill_marginals_matrix(marginals_matrix, data, arms, n_obs, compare_to_zero)
}
\arguments{
\item{marginals_matrix}{the marginals matrix in the dataframe format created previously}

\item{data}{the data with the dosage of each arm}

\item{arms}{a vector containing the names of each arm which are the arms columns names in "data"}

\item{n_obs}{the number of observations, basically the number of rows in "data"}

\item{compare_to_zero}{is a boolean,
if TRUE, a policy dominates a marginal if all dosages are greater
if FALSE, they must also have the exact same activated arms}
}
\value{
the marginals matrix with the right indicators inside
}
\description{
Fill the empty marginals matrix already crated
}
\examples{
arms = c('financial_incentive','reminder','information')
A1 = c(0,0,0,0,0,1,1,1,1,1)
A2 = c(1,1,0,0,1,1,0,0,1,1,0)
A3 = c(0,1,2,3,0,3,2,1,0,1)
data = data.frame(financial_incentive = A1, reminder = A2, information = A3)
n_obs = nrow(data)
empty_marginals_matrix = create_empty_marginals_matrix(c(1,1,3),n_obs)
fill_marginals_matrix(empty_marginals_matrix, data, arms, n_obs,FALSE)
}
