% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winners_curse.R
\name{get_hybrid_estimate}
\alias{get_hybrid_estimate}
\title{Winners curse}
\usage{
get_hybrid_estimate(first_Y, second_Y, first_Y_var, ntreat, alpha, beta)
}
\arguments{
\item{first_Y}{is the scaled best effect}

\item{first_Y_var}{is the sclaed best effect variation}

\item{ntreat}{is the number of estimated effects}

\item{alpha}{is the alpha used in the winners curse algorithm, default to 0.05}

\item{beta}{is the beta used in the winners curse algorithm}

\item{second}{Y is the sclaed second best effect}
}
\value{
returns a list with the median unbiased estimate of the best effect and the alpha confidence interval lower and upper bounds
}
\description{
Apply the winners curse algorithm to the final pooled OLS
}
