% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winners_curse.R
\name{get_hybrid_mu_alpha}
\alias{get_hybrid_mu_alpha}
\title{Compute the hybrid mu_alpha}
\usage{
get_hybrid_mu_alpha(first_Y, second_Y, c_beta, first_Y_var, alpha)
}
\arguments{
\item{first_Y}{is the scaled best effect}

\item{c_beta}{is the coefficient computed for the joint confidence interval of level beta}

\item{first_Y_var}{is the sclaed best effect variation}

\item{alpha}{is the alpha used in the winners curse algorithm, default to 0.05}

\item{second}{Y is the sclaed second best effect}
}
\value{
the solution mu_alpha
}
\description{
Solves the equation F_TN^H (Y(theta^hat),mu_alpha,theta^hat,Z_theta^hat) = 1-alpha
}
