% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{get_policy_fullname}
\alias{get_policy_fullname}
\title{Going from policy concise string to a full name}
\usage{
get_policy_fullname(policy, arms)
}
\arguments{
\item{policy}{a policy in the format c_A_B_C_D_.. where A,B,C,D,... are integers representing the dosage for arm 1,2,3,4,...}

\item{arms}{a vector containing the arms' names and of the same length as the vector representation of the policy}
}
\value{
the fullname of the policy with the dosage of each arm
}
\description{
Create the policy full name from its string representation
}
\examples{
get_policy_fullname('c_1_0_2', c('financial_incentive','reminder','information'))
}
