% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{get_pooled_ols}
\alias{get_pooled_ols}
\title{Perform the final OLS}
\usage{
get_pooled_ols(data, fes = c(), y, w = NULL, pool_ids, clusters)
}
\arguments{
\item{data}{is the dataframe containing all our data and pool_id dummy columns}

\item{fes}{is a vector containing the column names of all the fixed effects}

\item{y}{is the column name of the outcome of interest}

\item{w}{is the column name of the weights}

\item{pool_ids}{is a vector containing the column names of each pool_id dummy column}

\item{clusters}{(optional) is the column name that corresponds to the clusters in the data, that should be used in the final pooled OLS. Please refer to estimatr::lm_robust documentation for more information on this parameter.}
}
\value{
returns the ols result
}
\description{
Perform the final OLS on pooled data, regressing on fes and pool_id dummy columns
}
\examples{
arms = c('financial_incentive','reminder','information')
fes = c('fes_1')
y = 'outcome'
w = 'weights'
A1 = c(0,0,0,0,0,1,1,1,1,1)
A2 = c(1,1,0,0,1,1,0,0,1,1)
A3 = c(0,1,2,3,0,3,2,1,0,1)
F1 = c(0,1,0,0,0,1,0,1,0,0)
Y  = c(5,4,3,5,4,5,4,2,3,2)
W  = c(1,1,1,2,1,2,2,1,1,2)
p1 = c(0,1,0,0,0,0,0,0,0,0)
p2 = c(0,0,0,0,0,1,0,0,0,1)
pool_ids = c('pool_id_1','pool_id_2')
data = data.frame(pool_id_1 = p1, pool_id_2 = p2, financial_incentive = A1, reminder = A2, information = A3, fes_1 = F1, outcome = Y, weights=W)
get_pooled_ols(data,fes,y,w,pool_ids)
}
