% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{grid_pval_OSE}
\alias{grid_pval_OSE}
\title{P-values / size of support grid for multiple pval cutoffs in the pval one-step elimination}
\usage{
grid_pval_OSE(
  cutoffs = NULL,
  data,
  arms,
  fes = c(),
  y,
  w = NULL,
  scale = FALSE,
  compare_to_zero = FALSE
)
}
\arguments{
\item{cutoffs}{is a vector containing all the cutoffs we want to test in the grid, default to NULL\cr
if NULL, the cutoffs used is c(5e-1,4e-1,3e-1,2e-1, 1e-1, 5e-2, 1e-2, 5e-3, 1e-3, 1e-4, 1e-5, 1e-6, 1e-7, 1e-8, 1e-9, 1e-10, 1e-12, 1e-15, 1e-20, 1e-30)}

\item{data}{is the dataframe containing all our data}

\item{arms}{is a vector containing the column names of all the arms}

\item{fes}{is a vector containing the column names of all the fixed effects}

\item{y}{is the column name of the outcome of interest}

\item{w}{is the column name of the weights}

\item{scale}{is a boolean, if TRUE, the data is going to be scaled (mean 0 and sd 1), if FALSE, nothing happens}

\item{compare_to_zero}{is a boolean.\cr
If TRUE, the code considers that a policy dominates a marginal if all dosages are greater\cr
If FALSE, then they must also have the exact same activated arms (the zeros of the policy vectors are at identical indexes)}
}
\value{
returns a dataframe containing all the p-values in cutoffs and the according support size one would get for each of those p-values
}
\description{
Simulate many pval one-step elimination for different pval cutoffs values.\cr
This allows to see the size of support one would get for each pval cutoff.
}
\examples{
arms = c('financial_incentive','reminder','information')
fes = c('fes_1')
y = 'outcome'
w = 'weights'
A1 = c(0,0,0,0,0,1,1,1,1,1)
A2 = c(1,1,0,0,1,1,0,0,1,1)
A3 = c(0,1,2,3,0,3,2,1,0,1)
F1 = c(0,1,0,0,0,1,0,1,0,0)
Y  = c(5,4,3,5,4,5,4,2,3,2)
W  = c(1,1,1,2,1,2,2,1,1,2)
data = data.frame(financial_incentive = A1, reminder = A2, information = A3, fes_1 = F1, outcome = Y, weights=W)
grid_pval_OSE(cutoffs=NULL,data=data,arms=arms,fes=fes,y=y,w=w,scale=FALSE,compare_to_zero=FALSE)
}
