% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_estimation.R
\name{lambda_to_pval}
\alias{lambda_to_pval}
\title{From lambda to pval cutoff}
\usage{
lambda_to_pval(X, y, variables, lambda)
}
\arguments{
\item{X}{is the regression matrix with the marginals, the fixed effects, the outcome of interest}

\item{y}{is the outcome of interest}
}
\value{
the pval equivalent to lambda
}
\description{
Compute the equivalent p-value between Puffer_N LASSO lambda and pval one-step elimination
}
