% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_estimation.R
\name{lambda_to_threshold}
\alias{lambda_to_threshold}
\title{From lambda to threshold}
\usage{
lambda_to_threshold(X, y, variables, lambda)
}
\arguments{
\item{X}{is the regression matrix with the marginals, the fixed effects, the outcome of interest}

\item{y}{is the outcome of interest}
}
\value{
returns the threshold equivalent to lambda
}
\description{
Compute the equivalent threshold for the lambda of Puffer_N LASSO and pval one-step elimination
}
