% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_estimation.R
\name{N_transform}
\alias{N_transform}
\title{N transform}
\usage{
N_transform(X)
}
\arguments{
\item{X}{a matrix with the observations}
}
\value{
returns XN, where N transformation has been applied to X
}
\description{
Perform the N transformation on X and Y
}
