% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{pool_data}
\alias{pool_data}
\title{Pool the data}
\usage{
pool_data(data, arms, marginal_support_strings, compare_to_zero)
}
\arguments{
\item{data}{is the dataframe containing all our data}

\item{arms}{is a vector containing the column names of all the arms}

\item{marginal_support_strings}{is a vector containing strings that represent all the marginals in the support}

\item{compare_to_zero}{is a boolean.
If TRUE, the code considers that a policy dominates a marginal if all dosages are greater
If FALSE, then they must also have the exact same activated arms (the zeros of the policy vectors are at identical indexes)}
}
\value{
returns the dataframe "data" with a pool_id column that gives the id of the observation's pool, and one dummy column per pool
}
\description{
Take the estimated support and pool each observation in its right pool
}
\examples{
arms = c('financial_incentive','reminder','information')
A1 = c(0,0,0,0,0,1,1,1,1,1)
A2 = c(1,1,0,0,1,1,0,0,1,1,0)
A3 = c(0,1,2,3,0,3,2,1,0,1)
data = data.frame(financial_incentive = A1, reminder = A2, information = A3)
marginal_support_strings = c('c_0_1_1', 'c_1_1_2', 'c_1_0_1', 'c_1_0_2')
pool_data(data,arms,marginal_support_strings,FALSE)
}
