% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{pools_info}
\alias{pools_info}
\title{Give information for each pool}
\usage{
pools_info(data, arms)
}
\arguments{
\item{data}{is the dataframe containing all our data and a "pool_id" column giving the pool id of each observation}

\item{arms}{is a vector containing the column names of all the arms}
}
\value{
returns a list containing :\cr
\itemize{
\item pools_summary: a dataframe containing information on each pool, with columns:
\itemize{
\item "pool_id": the pool id
\item "n_unique_policies": the number of unique policies inside this pool
\item "n_obs": the number of observation in this pool
\item "pool_influences": a string of format "i_x1_x2_x3..", where x_i equals 1 if the marginal number i influences this pool, 0 otherwise
\item "pool_influences_list": a string that gives all the marginals that influence this pool
\item "pool_minimum_intensity": the smallest unique policy inside this pool
\item "min_arm1", "min_arm2" etc... : one column per arm, giving the minimum dosage for each arm inside this pool
\item "pool_minimum_intensity_fullname": the full name of the smallest unique policy
\item "examples_of_contained_policies": 5 or less examples of unique policies that are inside this pool
}
\item unique_policy: a dataframe containing information on each unique policy, with columns:
\itemize{
\item "policy": a string in format "p_A_B_C_..." where A is the dosage on arm 1, B on arm 2 etc... This is the definition of the unique policy
\item "policy_fullname": the full name of the unique policy
\item "arm1", "arm2" etc...: the value on each arm of the unique policy
\item "pool_id": the id of the pool the unique policy belongs to
\item "pool_influences": a string of format "i_x1_x2_x3..", where x_i equals 1 if the marginal number i influences this pool, 0 otherwise
\item "pool_influences_list": a string that gives all the marginals that influence this pool
}
}
}
\description{
Compute some useful information to understand what's inside each pool
}
\examples{
arms = c('financial_incentive','reminder','information')
A1 = c(0,0,0,0,0,1,1,1,1,1)
A2 = c(1,1,0,0,1,1,0,0,1,1)
A3 = c(0,1,2,3,0,3,2,1,0,1)
pool_id = c(0,1,0,0,0,2,0,0,0,2)
data = data.frame(financial_incentive = A1, reminder = A2, information = A3, pool_id = pool_id)
pools_info(data,arms)
}
