% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winners_curse.R
\name{ptmvnorm_extended}
\alias{ptmvnorm_extended}
\title{Extended truncated normal}
\usage{
ptmvnorm_extended(q, mu, sigma, lb, ub)
}
\arguments{
\item{q}{the point where we evaluate the function}

\item{mu}{the mean of the normal}

\item{sigma}{the standard deviation of the normal}

\item{lb}{the lower truncation point}

\item{ub}{the upper truncation point}
}
\value{
the extended truncated normal function
}
\description{
Extends the truncated normal functions in case the point is outside bounds
}
