% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_estimation.R
\name{puffer_transform}
\alias{puffer_transform}
\title{Puffer transform}
\usage{
puffer_transform(X, Y)
}
\arguments{
\item{X}{a matrix with the observations}

\item{Y}{a vector with the outcome of interest}
}
\value{
returns FX and FY, where puffer transformation has been applied to X and Y
}
\description{
Perform the puffer transformation on X and Y
}
