% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_estimation.R
\name{pval_MSE}
\alias{pval_MSE}
\title{pval multi-step elimination}
\usage{
pval_MSE(X, y, variables, threshold)
}
\arguments{
\item{X}{is the regression matrix with the marginals, the fixed effects, the outcome of interest}

\item{y}{is the outcome of interest}

\item{variables}{is a vector containing all the variables we want to regress on}

\item{threshold}{is the threshold on p-values}
}
\value{
returns the estimated support
}
\description{
Estimate support with p-value multi-step elimination
}
