% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_estimation.R
\name{pval_OSE}
\alias{pval_OSE}
\title{pval one-step elimination}
\usage{
pval_OSE(X, y, variables, pval_cutoff)
}
\arguments{
\item{X}{is the regression matrix with the marginals, the fixed effects, the outcome of interest}

\item{y}{is the outcome of interest}

\item{variables}{is a vector containing all the variables we want to regress on}

\item{pval_cutoff}{is the cutoff on p-values}
}
\value{
returns the estimated support
}
\description{
Estimate support with p-value one-step elimination
}
