% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_estimation.R
\name{pval_to_lambda}
\alias{pval_to_lambda}
\title{From pval to lambda cutoff}
\usage{
pval_to_lambda(X, y, variables, pval)
}
\arguments{
\item{X}{is the regression matrix with the marginals, the fixed effects, the outcome of interest}

\item{y}{is the outcome of interest}
}
\value{
returns the lambda equivalent to pval
}
\description{
Compute the equivalent lambda between Puffer_N LASSO lambda and pval one-step elimination
}
