% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{string_to_vector}
\alias{string_to_vector}
\title{Going from string to vector}
\usage{
string_to_vector(string)
}
\arguments{
\item{string}{is a string in the format 'c_A_B_C_D_..' where A,B,C,D,.. are integers}
}
\value{
returns a vector containing the values represented by the string, c(A,B,C,D,...)
}
\description{
Create a vector from its string representation
}
\examples{
string_to_vector('c_1_2_3_4_5')
}
