% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{suggest_pval_cutoff}
\alias{suggest_pval_cutoff}
\title{Suggest one p-value cutoff that could be used the in pval one-step elimination}
\usage{
suggest_pval_cutoff(
  data,
  arms,
  y,
  support_size_target = NULL,
  fes = c(),
  w = NULL,
  estim_func = "pval_OSE",
  compare_to_zero = FALSE,
  clusters = NULL
)
}
\arguments{
\item{data}{is the dataframe containing all our data}

\item{arms}{is a vector containing the column names of all the arms}

\item{y}{is the column name of the outcome of interest}

\item{support_size_target}{(optional) is the prior we might have on the number of marginals in the support. It is NULL by default. \cr
If NULL, the code will take the support size that corresponds to the elbow of the R-squared vs support size curve.}

\item{fes}{is a vector containing the column names of all the fixed effects}

\item{w}{is the column name of the weights}

\item{compare_to_zero}{is a boolean.\cr
If TRUE, the code considers that a policy dominates a marginal if all dosages are greater\cr
If FALSE, then they must also have the exact same activated arms (the zeros of the policy vectors are at identical indexes)}
}
\value{
returns a pval cutoff suggestion with the according number of pools it will produce
}
\description{
Suggest one p-value cutoff that could be used the in pval one-step elimination.\cr
This is done in two steps. First, we take the p-value corresponding to a target, which can be provided by the user, or which corresponds to the elbow of the R-squared vs support size curve. \cr
Then, we look for p-values arround this target and check if some of them produce a final pooled OLS with two best effects that are statistically different from 0.
}
\examples{
arms = c('financial_incentive','reminder','information')
fes = c('fes_1')
y = 'outcome'
w = 'weights'
A1 = c(0,0,0,0,0,1,1,1,1,1)
A2 = c(1,1,0,0,1,1,0,0,1,1)
A3 = c(0,1,2,3,0,3,2,1,0,1)
F1 = c(0,1,0,0,0,1,0,1,0,0)
Y  = c(5,4,3,5,4,5,4,2,3,2)
W  = c(1,1,1,2,1,2,2,1,1,2)
data = data.frame(financial_incentive = A1, reminder = A2, information = A3, fes_1 = F1, outcome = Y, weights=W)
suggest_pval_OSE_cutoff(data=data,arms=arms,y=y,fes=fes,w=w,compare_to_zero=FALSE)
}
