% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{suggest_pval_OSE_cutoff}
\alias{suggest_pval_OSE_cutoff}
\title{Suggest one p-value cutoff that could be used the in pval one-step elimination}
\usage{
suggest_pval_OSE_cutoff(
  data,
  arms,
  fes = c(),
  y,
  w = NULL,
  scale = FALSE,
  compare_to_zero = FALSE
)
}
\arguments{
\item{data}{is the dataframe containing all our data}

\item{arms}{is a vector containing the column names of all the arms}

\item{fes}{is a vector containing the column names of all the fixed effects}

\item{y}{is the column name of the outcome of interest}

\item{w}{is the column name of the weights}

\item{scale}{is a boolean, if TRUE, the data is going to be scaled (mean 0 and sd 1), if FALSE, nothing happens}

\item{compare_to_zero}{is a boolean.\cr
If TRUE, the code considers that a policy dominates a marginal if all dosages are greater\cr
If FALSE, then they must also have the exact same activated arms (the zeros of the policy vectors are at identical indexes)}
}
\value{
returns a pval cutoff suggestion with the according number of pools it will produce
}
\description{
Suggest one p-value cutoff that could be used the in pval one-step elimination.\cr
This is done by computing the grid_pval_OSE and taking the pval cutoff that create a number of pools equal to the number of unique policies divided by 20.
}
\examples{
arms = c('financial_incentive','reminder','information')
fes = c('fes_1')
y = 'outcome'
w = 'weights'
A1 = c(0,0,0,0,0,1,1,1,1,1)
A2 = c(1,1,0,0,1,1,0,0,1,1)
A3 = c(0,1,2,3,0,3,2,1,0,1)
F1 = c(0,1,0,0,0,1,0,1,0,0)
Y  = c(5,4,3,5,4,5,4,2,3,2)
W  = c(1,1,1,2,1,2,2,1,1,2)
data = data.frame(financial_incentive = A1, reminder = A2, information = A3, fes_1 = F1, outcome = Y, weights=W)
suggest_pval_OSE_cutoff(data=data,arms=arms,fes=fes,y=y,w=w,scale=FALSE,compare_to_zero=FALSE)
}
