% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_estimation.R
\name{threshold_to_lambda}
\alias{threshold_to_lambda}
\title{From threshold to lambda}
\usage{
threshold_to_lambda(X, y, variables, pval)
}
\arguments{
\item{X}{is the regression matrix with the marginals, the fixed effects, the outcome of interest}

\item{y}{is the outcome of interest}
}
\value{
returns the lambda equivalent to pval
}
\description{
Compute the equivalent lambda between Puffer_N LASSO and threshold in pval one-step elimination
}
