% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{do_TVA}
\alias{do_TVA}
\title{TVA Main function}
\usage{
do_TVA(
  data,
  arms,
  fes = c(),
  y,
  w = NULL,
  cutoff,
  estimation_function_name = "pval_OSE",
  scale = FALSE,
  compare_to_zero = FALSE
)
}
\arguments{
\item{data}{is the dataframe containing all our data}

\item{arms}{is a vector containing the column names of all the arms}

\item{fes}{is a vector containing the column names of all the fixed effects}

\item{y}{is the column name of the outcome of interest}

\item{w}{is the column name of the weights}

\item{cutoff}{is the cutoff used in the support estimation}

\item{estimation_function_name}{is the estimation function we should used. Possible functions are :\cr
\enumerate{
\item pval_MSE: a multiple step elimination on p-values\cr
\item pval_OSE: a one step elimination on p-values\cr
\item puffer_N_LASSO: a LASSO OLS with a Puffer_N transformation\cr
\item beta_OSE: a one step elimination on beta values\cr
\item puffer_LASSO: a LASSO OLS with a Puffer transformation\cr
(2) and (3) should be equivalent, as well as (4) and (5)
}}

\item{scale}{is a boolean, if TRUE, the data is going to be scaled (mean 0 and sd 1), if FALSE, nothing happens}

\item{compare_to_zero}{is a boolean.\cr
If TRUE, the code considers that a policy dominates a marginal if all dosages are greater\cr
If FALSE, then they must also have the exact same activated arms (the zeros of the policy vectors are at identical indexes)}
}
\value{
returns a list containing:\cr
data: the data with new columns giving pooling information\cr
marginal_support: a dataframe with all the marginals in the support and their according id\cr
pools_summary: a dataframe with information on each pool\cr
unique_policy: a dataframe with all the possible unique policies and their according pool id\cr
fes_support: the intersection between the estimated support and the fixed effects\cr
pooled_ols: the result of the final OLS on the pooled data\cr
winners_effect: the result of the best pooled policy effect, downsized by the winners curse algorithm
}
\description{
Perform the whole TVA algorithm
}
\examples{
arms = c('financial_incentive','reminder','information')
fes = c('fes_1')
y = 'outcome'
w = 'weights'
A1 = c(0,0,0,0,0,1,1,1,1,1)
A2 = c(1,1,0,0,1,1,0,0,1,1,0)
A3 = c(0,1,2,3,0,3,2,1,0,1)
F1 = c(0,1,0,0,0,1,0,1,0,0)
Y  = c(5,4,3,5,4,5,4,2,3,2)
W  = c(1,1,1,2,1,2,2,1,1,2)
data = data.frame(financial_incentive = A1, reminder = A2, information = A3, fes_1 = F1, outcome = Y, weights=W)
TVA(data,arms,fes,y,w,0.3,'pval_OSE',FALSE,FALSE)
}
