% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{fill_marginals_matrix}
\alias{fill_marginals_matrix}
\title{Fill marginals matrix}
\usage{
fill_marginals_matrix(marginals_matrix, data, arms, n_obs, compare_to_zero)
}
\arguments{
\item{marginals_matrix}{is the marginals matrix in the dataframe format created previously}

\item{data}{is the data with the dosage on each arm}

\item{arms}{is a vector containing the column names that represent the dosages of each arm in data}

\item{n_obs}{is the number of observations, equal to the number of rows in "data"}

\item{compare_to_zero}{is a boolean.\cr
If TRUE, the code considers that a policy dominates a marginal if all dosages are greater\cr
If FALSE, then they must also have the exact same activated arms (the zeros of the policy vectors are at identical indexes)}
}
\value{
returns the marginals matrix with the right indicators inside
}
\description{
Fill the empty marginals matrix already created
}
\examples{
arms = c('financial_incentive','reminder','information')
A1 = c(0,0,0,0,0,1,1,1,1,1)
A2 = c(1,1,0,0,1,1,0,0,1,1)
A3 = c(0,1,2,3,0,3,2,1,0,1)
data = data.frame(financial_incentive = A1, reminder = A2, information = A3)
n_obs = nrow(data)
empty_marginals_matrix = create_empty_marginals_matrix(c(1,1,3),n_obs)
fill_marginals_matrix(empty_marginals_matrix, data, arms, n_obs,FALSE)
}
