% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{pools_info}
\alias{pools_info}
\title{Give information for each pool}
\usage{
pools_info(data, arms)
}
\arguments{
\item{data}{is the dataframe containing all our data and a "pool_id" column giving the pool id of each observation}

\item{arms}{is a vector containing the column names of all the arms}
}
\value{
a list containing :
pools_summary: a dataframe containing information on each pool (number of unique policies, number of observations, pool minimum etc...)
unique_policy: a dataframe containing all the possible unique policies with their according pool id
}
\description{
Compute some useful information to understand what's inside each pool
}
\examples{
arms = c('financial_incentive','reminder','information')
A1 = c(0,0,0,0,0,1,1,1,1,1)
A2 = c(1,1,0,0,1,1,0,0,1,1,0)
A3 = c(0,1,2,3,0,3,2,1,0,1)
pool_id = c(0,1,0,0,0,2,0,0,0,2)
data = data.frame(financial_incentive = A1, reminder = A2, information = A3, pool_id = pool_id)
pools_info(data,arms)
}
