% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare the data}
\usage{
prepare_data(data, arms, y, fes, w, compare_to_zero)
}
\arguments{
\item{data}{is the dataframe containing all our data}

\item{arms}{is a vector containing the column names of all the arms}

\item{y}{is the column name of the outcome of interest}

\item{fes}{is a vector containing the column names of all the fixed effects}

\item{w}{is the column name of the weights}

\item{compare_to_zero}{is a boolean.\cr
If TRUE, the code considers that a policy dominates a marginal if all dosages are greater\cr
If FALSE, then they must also have the exact same activated arms (the zeros of the policy vectors are at identical indexes)}
}
\value{
returns a list containing :\cr
\itemize{
\item X: a dataframe containing the marginals matrix, the fixed effects, the outcome and the intercept. X is ready for support estimation\cr
\item data: the original dataframe with new columns for arm dosages (if dummy arm columns were provided by the user)
\item arms: a vector containing the columns with the dosages on each arms (no more dummy columns if it was provided by the user). \cr
\item marginals_colnames: a vector containing all the marginals names (also called the alphas)\cr
\item variables: the list of variables that should be used in the regression (fixed effects, marginals and the intercept)
}
}
\description{
Prepare the data for the support estimation
}
\examples{
arms = c('financial_incentive','reminder','information')
fes = c('fes_1')
y = 'outcome'
w = 'weights'
A1 = c(0,0,0,0,0,1,1,1,1,1)
A2 = c(1,1,0,0,1,1,0,0,1,1)
A3 = c(0,1,2,3,0,3,2,1,0,1)
F1 = c(0,1,0,0,0,1,0,1,0,0)
Y  = c(5,4,3,5,4,5,4,2,3,2)
W  = c(1,1,1,2,1,2,2,1,1,2)
data = data.frame(financial_incentive = A1, reminder = A2, information = A3, fes_1 = F1, outcome = Y, weights=W)
prepare_data(data,arms,y, fes,w,FALSE,FALSE)
}
