% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_estimation.R
\name{puffer_LASSO}
\alias{puffer_LASSO}
\title{Puffer LASSO}
\usage{
puffer_LASSO(X, y, variables, lambda)
}
\arguments{
\item{X}{is the regression matrix with the marginals, the fixed effects, the outcome of interest}

\item{y}{is the outcome of interest}

\item{variables}{is a vector containing all the variables we want to regress on}

\item{lambda}{is the cutoff on beta values}
}
\value{
returns the estimated support
}
\description{
Estimate support with Puffer LASSO
}
