% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{vector_to_string}
\alias{vector_to_string}
\title{Going from vector to string}
\usage{
vector_to_string(vector)
}
\arguments{
\item{vector}{is a vector of any size and containing integers}
}
\value{
returns a string representing the vector in format 'c_A_B_C_D_...' where A,B,C,D,.. are the integers in the vector
}
\description{
Create a string representing the vector
}
\examples{
vector_to_string(c(1,2,3,4,5))
vector_to_string(c(10))
}
